feat: Refactored the main stats table into sub-components
