+++ client/src/OpeningStatsTable.js
@@ -1,8 +1,6 @@
 import React, { useEffect, useState, useMemo } from "react";
-import Select, { components } from "react-select";
-import "react-datepicker/dist/react-datepicker.css";
-import { generateChessComLink } from "./chessLinkUtils";
-import DateRangeSelector from "./DateRangeSelector";
+import SummaryBar from "./SummaryBar";
+import OpeningTable from "./OpeningTable";
 
 const OpeningStatsTable = ({
   data = {},
@@ -13,7 +11,7 @@ const OpeningStatsTable = ({
   fullResetTrigger,
   color = null,
   summaryLabel = "All Games",
-  testId = `Openings filter- {color}`,
+  testId = `Openings filter- ${color}`,
 }) => {
   const [sortColumn, setSortColumn] = useState("played");
   const [sortOrder] = useState("desc");
@@ -23,20 +21,8 @@ const OpeningStatsTable = ({
   const [showingFilteredSummary, setShowingFilteredSummary] = useState(false);
   const [dateRangeOption, setDateRangeOption] = useState("last-30");
 
-  const CustomControl = ({ children, innerRef, innerProps, ...rest }) => (
-    <components.Control
-      {...rest}
-      innerRef={innerRef}
-      innerProps={{ ...innerProps, "data-test-id": testId }}
-    >
-      {children}
-    </components.Control>
-  );
-
   useEffect(() => {
-    if (resetToDefaultRange) {
-      setDateRangeOption("last-30");
-    }
+    if (resetToDefaultRange) setDateRangeOption("last-30");
   }, [resetToDefaultRange]);
 
   useEffect(() => {
@@ -78,11 +64,6 @@ const OpeningStatsTable = ({
     return Object.fromEntries(sorted.slice(0, viewLimit));
   }, [data, selectedOptions, sortColumn, sortOrder, viewLimit]);
 
-  const clearFilters = () => {
-    setSelectedOptions([]);
-    setViewLimit(5);
-  };
-
   const fullSummary = useMemo(() => {
     return Object.entries(data).reduce(
       (acc, [key, val]) => {
@@ -110,36 +91,6 @@ const OpeningStatsTable = ({
     );
   }, [filteredEntries]);
 
-  const MultiValue = ({ index, getValue, ...props }) => {
-    const maxToShow = 3;
-    if (index < maxToShow) return <components.MultiValue {...props} />;
-    if (index === maxToShow) {
-      return (
-        <components.MultiValue {...props}>
-          <span style={{ paddingLeft: "4px" }}>...</span>
-        </components.MultiValue>
-      );
-    }
-    return null;
-  };
-
-  const renderLinkedCell = (count, resultType, ecoUrlString) => {
-    if (count > 0 && data.startDate && data.endDate) {
-      const startFormatted = new Date(data.startDate).toLocaleDateString(
-        "en-US",
-      );
-      const endFormatted = new Date(data.endDate).toLocaleDateString("en-US");
-      const baseLink = generateChessComLink(ecoUrlString, resultType, color);
-      const dateQuery = `&endDate%5Bdate%5D=${encodeURIComponent(endFormatted)}&startDate%5Bdate%5D=${encodeURIComponent(startFormatted)}`;
-      return (
-        <a href={baseLink + dateQuery} target="_blank" rel="noreferrer">
-          {count}
-        </a>
-      );
-    }
-    return count;
-  };
-
   const renderSortIndicator = (column) => {
     if (column !== sortColumn) return null;
     return sortOrder === "asc" ? " ↑" : " ↓";
@@ -155,150 +106,37 @@ const OpeningStatsTable = ({
 
   return (
     <div data-test-id={`Openings table - ${color || "both"}`}>
-      <div className="summary-bar-vertical">
-        <div className="summary-text">
-          {showingFilteredSummary ? (
-            <>
-              Filtered {summaryLabel}: (P: {filteredSummary.played}, W:{" "}
-              {filteredSummary.won}, L: {filteredSummary.lost}, D:{" "}
-              {filteredSummary.drawn})<br />
-              {summaryLabel}: (P: {fullSummary.played}, W: {fullSummary.won}, L:{" "}
-              {fullSummary.lost}, D: {fullSummary.drawn})<br />
-            </>
-          ) : (
-            <>
-              {summaryLabel}: (P: {fullSummary.played}, W: {fullSummary.won}, L:{" "}
-              {fullSummary.lost}, D: {fullSummary.drawn})<br />
-            </>
-          )}
-          Showing {visibleRows} rows out of {totalRows}
-        </div>
-
-        <div className="filter-dropdown">
-          <Select
-            options={filterOptions}
-            isMulti
-            placeholder="Filter openings"
-            value={selectedOptions}
-            onChange={(newOptions, actionMeta) => {
-              if (
-                ["select-option", "remove-value"].includes(actionMeta.action)
-              ) {
-                setShowingFilteredSummary(true);
-              } else if (actionMeta.action === "clear") {
-                setShowingFilteredSummary(false);
-              }
-              setSelectedOptions(newOptions || []);
-            }}
-            classNamePrefix="react-select"
-            components={{ MultiValue, Control: CustomControl }}
-            styles={{
-              control: (base) => ({
-                ...base,
-                minHeight: "30px",
-                fontSize: "12px",
-              }),
-            }}
-            data-test-id={testId}
-          />
-        </div>
-
-        <div className="date-range-dropdown">
-          <DateRangeSelector
-            onDateRangeResolved={onDateRangeChange}
-            dateRangeOption={dateRangeOption}
-            setDateRangeOption={setDateRangeOption}
-            testId={`date range selector - ${color || "all"}`}
-          />
-        </div>
-
-        <div className="summary-buttons">
-          <button
-            onClick={() => {
-              setViewLimit(5);
-              setShowingFilteredSummary(true);
-            }}
-          >
-            Show 5
-          </button>
-          <button
-            onClick={() => {
-              setViewLimit(10);
-              setShowingFilteredSummary(true);
-            }}
-          >
-            Show 10
-          </button>
-          <button
-            onClick={() => {
-              setViewLimit(totalRows);
-              setShowingFilteredSummary(selectedOptions.length > 0);
-            }}
-          >
-            Show All
-          </button>
-          <button
-            onClick={() => {
-              clearFilters();
-              if (onResetToCachedOneMonth) onResetToCachedOneMonth();
-              setShowingFilteredSummary(false);
-            }}
-          >
-            Clear Filters
-          </button>
-        </div>
-      </div>
-
-      {/* Table section */}
-      <div className="table-wrapper" style={{ position: "relative" }}>
-        {loading && <div className="loading-overlay">Loading data...</div>}
-        <table className="chess-table" style={{ opacity: loading ? 0.5 : 1 }}>
-          <thead>
-            <tr data-test-id={`Table header for ${color || "all"} openings`}>
-              <th className="sortable" onClick={() => setSortColumn("name")}>
-                Opening Name{renderSortIndicator("name")}
-              </th>
-              <th>Opening Name with Eco</th>
-              <th className="sortable" onClick={() => setSortColumn("played")}>
-                Played{renderSortIndicator("played")}
-              </th>
-              <th className="sortable" onClick={() => setSortColumn("won")}>
-                Won{renderSortIndicator("won")}
-              </th>
-              <th className="sortable" onClick={() => setSortColumn("lost")}>
-                Lost{renderSortIndicator("lost")}
-              </th>
-              <th className="sortable" onClick={() => setSortColumn("drawn")}>
-                Drawn{renderSortIndicator("drawn")}
-              </th>
-            </tr>
-          </thead>
-          <tbody>
-            {Object.entries(filteredEntries).map(([name, stats]) => (
-              <tr key={name} data-test-id="Chess data row">
-                <td>{name}</td>
-                <td>
-                  <a href={stats.ecoUrl} target="_blank" rel="noreferrer">
-                    {stats.ecoCode}
-                  </a>
-                </td>
-                <td>
-                  {renderLinkedCell(stats.played, null, stats.ecoUrlString)}
-                </td>
-                <td>
-                  {renderLinkedCell(stats.won, "win", stats.ecoUrlString)}
-                </td>
-                <td>
-                  {renderLinkedCell(stats.lost, "lost", stats.ecoUrlString)}
-                </td>
-                <td>
-                  {renderLinkedCell(stats.drawn, "draw", stats.ecoUrlString)}
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
+      <SummaryBar
+        {...{
+          summaryLabel,
+          fullSummary,
+          filteredSummary,
+          selectedOptions,
+          setSelectedOptions,
+          filterOptions,
+          showingFilteredSummary,
+          setShowingFilteredSummary,
+          dateRangeOption,
+          setDateRangeOption,
+          onDateRangeChange,
+          onResetToCachedOneMonth,
+          viewLimit,
+          setViewLimit,
+          totalRows,
+          visibleRows,
+          testId,
+          color,
+        }}
+      />
+      <OpeningTable
+        color={color}
+        loading={loading}
+        filteredEntries={filteredEntries}
+        renderSortIndicator={renderSortIndicator}
+        setSortColumn={setSortColumn}
+        startDate={data.startDate}
+        endDate={data.endDate}
+      />
     </div>
   );
 };
